Ash's first commit with all tasks
- Add display_name to product slot model
- Donation page - no decimal portion
- Rework index page
- Remove extra stuff from nav bar
- Create profile menu with icon from UX designer
- Added '$' to Donation Page

+++ hifireg/registration/forms/registration.py
@@ -109,6 +109,6 @@ class RegCompCodeForm(forms.Form):
 
 
 class RegDonateForm(forms.Form):
-    donation = forms.DecimalField(
-        label='I would like to contribute this dollar amount:', initial=0.00, decimal_places=2
+    donation = forms.IntegerField(
+        label='I would like to contribute this dollar amount:', initial=0
     )
